M11_Floristeria   NIVEL 2 	**Arregla menú retirar productos

+++ M11_PairProgramming_Floristeria_N2/src/com/nivel2/controller/removeProdMenu/RemoveFlowerController.java
@@ -13,46 +13,41 @@ import com.nivel2.view.ReadInfoWindow;
 import com.nivel2.view.ShowInfoWindow;
 import com.nivel2.view.utils.Session;
 
-public class RemoveFlowerController  extends Controller{
+public class RemoveFlowerController extends Controller {
 
 	FloristRepository floristRepository;
 
-	
 	public RemoveFlowerController(Session session) {
 		super(session);
 		this.floristRepository = FloristRepository.instance();
 	}
-	
+
 	public void control() {
 		int activeFloristId = ActiveFlorist.instance().getId();
-		ShowInfoWindow.showInfo(floristRepository.getProducts(activeFloristId)
-				.stream().filter(s -> s instanceof Flower).map(l -> l.toString()).collect(Collectors.toList()));
-		
-		
-		Product product = floristRepository.getProductById(activeFloristId, this.getProdId());
-		if(product instanceof Flower) {
-			this.floristRepository.remove(product, activeFloristId);
-			ShowInfoWindow.showInfo("FLOR ELIMINADA");
-		}		
-//		}else {
-//			ShowInfoWindow.showInfo("ERROR - ESTE ID NO ES DE UNA FLOR");
-//		}
-		
+	//**************************HE PUESTO ESTO PARA AVISAR DE QUE EL STOCK ESTÁ VACÍO*************
+		List<String> flowerStock = floristRepository.getProducts(activeFloristId).stream()
+				.filter(s -> s instanceof Flower).map(l -> l.toString()).collect(Collectors.toList());
+		ShowInfoWindow.showInfo(flowerStock);
+		if (flowerStock.isEmpty()) {
+			ShowInfoWindow.showInfo("NO HAY PRODUCTOS PARA RETIRAR DE ESTE STOCK");
+		} else {
+	//**************HASTA AQUÍ ES NUEVO**********
+			Product product = floristRepository.getProductById(activeFloristId, this.getProdId());
+			if (product instanceof Flower) {
+				this.floristRepository.remove(product, activeFloristId);
+				ShowInfoWindow.showInfo("FLOR ELIMINADA");
+			}
+		}
 	}
-	
-//	private int getProdId() {
-//		return ReadInfoWindow.readId(MessageView.CHOOSE_PRODUCT, ActiveFlorist.instance().getProducts().size());
-//	}
-	
+
 	private int getProdId() {
 		return ReadInfoWindow.readIdProduct(MessageView.CHOOSE_PRODUCT, this.getIdList());
 	}
-	
+
 	private List<Integer> getIdList() {
 		int activeFloristId = ActiveFlorist.instance().getId();
-		return this.floristRepository.getProducts(activeFloristId).stream().filter(s-> s instanceof Flower).map(s -> s.getId()).collect(Collectors.toList());
+		return this.floristRepository.getProducts(activeFloristId).stream().filter(s -> s instanceof Flower)
+				.map(s -> s.getId()).collect(Collectors.toList());
 	}
-	
-	
-}
 
+}
